// ARCHIVE_SINAMIND/sinamind_ai_analysis_dossier/frontend/src/lib/types.ts
export interface Chapter {
  id: string
  numero: number
  titre_fr: string
  titre_ar?: string
  page?: number
  type?: ChapterType
  importance: ChapterImportance
}

export interface Unit {
  id: string
  numero: number
  titre_fr: string
  titre_ar?: string
  page?: number
  chapters: Chapter[]
}

export interface Domain {
  id: string
  numero: number
  titre_fr: string
  titre_ar?: string
  units: Unit[]
}